feat(desktop): W7.1 extract DesktopDispatcher from ChatDelegate (byte-equivalent)
Move the 634-line ChatDelegate::execute_tool_calls body into a new
DesktopDispatcher<'a>::dispatch in desktop-client/ironclaw/src/agent/
desktop_dispatcher.rs. The original site becomes a 19-line stub that
forwards into DesktopDispatcher. Behavior unchanged.

W7.1 of ADR-160 §3 (L2 ToolDispatcher landing — desktop downstream).
W7.2 will impl dasclaw_runtime::ToolDispatcher on DesktopDispatcher.
W7.3 will wire it through AgenticLoop.

Helpers bumped to pub(super) so the new module can call them:
  - is_tool_disabled_by_extensions
  - PreflightOutcome
  - preflight_rejection_tool_message
  - contextual_tool_message

Two now-dead imports removed from dispatcher.rs:
  - tokio::task::JoinSet
  - crate::tools::redact_params

Single intentional deviation from a literal byte-for-byte port:
the plan-mode loop call site passes self.disabled_extensions
(not &self.disabled_extensions) because the new struct already
holds it as &'a HashSet<String>.

Cross-cuts: 类A (no new .ironclaw / IRONCLAW_BASE_DIR literals).

Verification:
  - cargo check -p dasclaw --tests : clean
  - cargo nextest run -p dasclaw -E 'test(dispatcher) | test(agentic_loop) | test(execute_tool) | test(desktop_dispatcher)' : 79/79 passed
  - cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings : clean
  - cargo fmt --all : clean
  - python3.12 scripts/check_no_panics.py --base origin/xClaw : clean

Closes #967
